requirements_local.txt
openai-whisper>=20240930
torch>=2.0.0
torchaudio>=2.0.0
opencv-python>=4.8.0
easyocr>=1.7.0
Pillow>=10.0.0
numpy>=1.24.0
